+++ bambi/families/univariate.py
@@ -29,25 +29,19 @@ class UnivariateFamily(Family):
 class Bernoulli(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "logit", "probit", "cloglog"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         "Sample from posterior predictive distribution"
         mean = self.link.linkinv(linear_predictor)
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
         return np.random.binomial(1, mean)
 
 
 class Beta(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "logit", "probit", "cloglog"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         kappa = posterior[model.response.name + "_kappa"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        kappa = kappa[:, idxs, np.newaxis]
-
+        kappa = kappa[:, :, np.newaxis]
         alpha = mean * kappa
         beta = (1 - mean) * kappa
         return np.random.beta(alpha, beta)
@@ -56,26 +50,20 @@ class Beta(UnivariateFamily):
 class Binomial(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "logit", "probit", "cloglog"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n, trials=None):
+    def posterior_predictive(self, model, posterior, linear_predictor, trials=None):
         if trials is None:
             trials = model.response.data[:, 1]
-
         mean = self.link.linkinv(linear_predictor)
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
         return np.random.binomial(trials.squeeze(), mean)
 
 
 class Gamma(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "log", "inverse"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         alpha = posterior[model.response.name + "_alpha"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        alpha = alpha[:, idxs, np.newaxis]
+        alpha = alpha[:, :, np.newaxis]
         beta = alpha / mean
         return np.random.gamma(alpha, 1 / beta)
 
@@ -83,58 +71,45 @@ class Gamma(UnivariateFamily):
 class Gaussian(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "log", "inverse"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         "Sample from posterior predictive distribution"
         mean = self.link.linkinv(linear_predictor)
         sigma = posterior[model.response.name + "_sigma"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        sigma = sigma[:, idxs, np.newaxis]
-
+        sigma = sigma[:, :, np.newaxis]
         return np.random.normal(mean, sigma)
 
 
 class NegativeBinomial(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "log", "cloglog"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         n = posterior[model.response.name + "_alpha"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        n = n[:, idxs, np.newaxis]
+        n = n[:, :, np.newaxis]
         p = n / (mean + n)
-
         return np.random.negative_binomial(n, p)
 
 
 class Poisson(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "log"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
         return np.random.poisson(mean)
 
 
 class StudentT(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "log", "inverse"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         sigma = posterior[model.response.name + "_sigma"].values
+        sigma = sigma[:, :, np.newaxis]
 
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
         if isinstance(self.likelihood.priors["nu"], (int, float)):
             nu = self.likelihood.priors["nu"]
         else:
-            nu = posterior[model.response.name + "_nu"].values[:, idxs, np.newaxis]
-
-        mean = mean[:, idxs, :]
-        sigma = sigma[:, idxs, np.newaxis]
+            nu = posterior[model.response.name + "_nu"].values[:, :, np.newaxis]
 
         return stats.t.rvs(nu, mean, sigma)
 
@@ -142,26 +117,18 @@ class StudentT(UnivariateFamily):
 class VonMises(UnivariateFamily):
     SUPPORTED_LINKS = ["identity", "tan_2"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         kappa = posterior[model.response.name + "_kappa"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        kappa = kappa[:, idxs, np.newaxis]
-
+        kappa = kappa[:, :, np.newaxis]
         return np.random.vonmises(mean, kappa)
 
 
 class Wald(UnivariateFamily):
     SUPPORTED_LINKS = ["inverse", "inverse_squared", "identity", "log"]
 
-    def posterior_predictive(self, model, posterior, linear_predictor, draws, draw_n):
+    def posterior_predictive(self, model, posterior, linear_predictor):
         mean = self.link.linkinv(linear_predictor)
         lam = posterior[model.response.name + "_lam"].values
-
-        idxs = np.random.randint(low=0, high=draw_n, size=draws)
-        mean = mean[:, idxs, :]
-        lam = lam[:, idxs, np.newaxis]
-
+        lam = lam[:, :, np.newaxis]
         return np.random.wald(mean, lam)